Server/requirements.txt
# Core ASGI and WebSocket
uvicorn==0.24.0
starlette==0.32.0
asgiref==3.7.2
fastapi==0.104.1

# Audio processing and VAD
webrtcvad==2.0.10
numpy>=1.24.3

# AI and ML - Updated for Gemini multimodal
google-generativeai==0.8.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
